feat: Move httpx to an optional extra for post-process support. (#17)
`pip install changelog-gen[post-process]`

+++ changelog_gen/version.py
@@ -42,7 +42,7 @@ class ModifyFile:
     path: Path
     patterns: list[str]
 
-    def update(self: t.Self, current: str, new: str) -> tuple[Path, Path]:
+    def update(self: t.Self, current: str, new: str, *, dry_run: bool) -> tuple[Path, Path]:
         """Update file with configured patterns."""
         with self.path.open("r") as f:
             contents = f.read()
@@ -57,8 +57,9 @@ class ModifyFile:
             contents = contents.replace(search, replace)
 
         backup = Path(f"{self.path}.bak")
-        with backup.open("w") as f:
-            f.write(contents)
+        if not dry_run:
+            with backup.open("w") as f:
+                f.write(contents)
 
         return (self.path, backup)
 
@@ -135,7 +136,7 @@ class BumpVersion:  # noqa: D101
         }
 
     @timer
-    def replace(self: T, current: Version, version: Version) -> list[str]:  # noqa: D102
+    def replace(self: T, current: Version, version: Version) -> list[str]:  # noqa: D102, C901
         if self.config.current_version == "":
             try:
                 find_config_file  # noqa: B018
@@ -179,14 +180,16 @@ class BumpVersion:  # noqa: D101
         modified_files = []
         for file in files_to_modify.values():
             try:
-                modified_files.append(file.update(current.raw, version.raw))
+                modified_files.append(file.update(current.raw, version.raw, dry_run=self.dry_run))
             except Exception:  # noqa: PERF203
-                for _original, backup in modified_files:
-                    backup.unlink()
+                if not self.dry_run:
+                    for _original, backup in modified_files:
+                        backup.unlink()
                 raise
 
-        for original, backup in modified_files:
-            original.write_text(backup.read_text())
-            backup.unlink()
+        if not self.dry_run:
+            for original, backup in modified_files:
+                original.write_text(backup.read_text())
+                backup.unlink()
 
         return sorted({mf.filename for mf in files_to_modify.values()})
